add mercurial and refactor parrot
