+++ projects/helper/solana.js
@@ -1,0 +1,23 @@
+const axios = require('axios')
+
+async function getTokenBalance(token, account) {
+    const tokenBalance = await axios.post("https://api.mainnet-beta.solana.com", {
+        "jsonrpc": "2.0",
+        "id": 1,
+        "method": "getTokenAccountsByOwner",
+        "params": [
+            account,
+            {
+                "mint": token
+            },
+            {
+                "encoding": "jsonParsed"
+            }
+        ]
+    })
+    return tokenBalance.data.result.value[0].account.data.parsed.info.tokenAmount.uiAmount
+}
+
+module.exports = {
+    getTokenBalance
+}
