+++ projects/parrot.js
@@ -1,24 +1,4 @@
-const axios = require('axios')
-
-"DGi3TxcKUq3E5t1mL33n9jRgdWWKngeRkP3fUppG4inn"
-
-async function getTokenBalance(token, account) {
-    const tokenBalance = await axios.post("https://api.mainnet-beta.solana.com", {
-        "jsonrpc": "2.0",
-        "id": 1,
-        "method": "getTokenAccountsByOwner",
-        "params": [
-            account,
-            {
-                "mint": token
-            },
-            {
-                "encoding": "jsonParsed"
-            }
-        ]
-    })
-    return tokenBalance.data.result.value[0].account.data.parsed.info.tokenAmount.uiAmount
-}
+const {getTokenBalance} = require('./helper/solana')
 
 async function tvl() {
     const [usdcAmount, usdtAmount] = await Promise.all([
